beta release of v2 docker-nas

+++ domain-add.sh
@@ -1,17 +1,11 @@
 #!/bin/bash
 
-#root check
-if [[ $EUID -ne 0 ]]; then
-   echo "This script must be run as root" 
-   exit 1
-fi
-
 cd "$(dirname "$0")"
 
-#load env variables
-source ./env.sh
+# Load env variables
+source .env
 
-#check if nginx is running
+# Check if nginx is running
 DOCKER_STOPPED=false
 if [ "$(docker ps -q -f name=nginx)" ]; then
     if [ "$(docker ps -aq -f status=running -f name=nginx)" ]; then
@@ -21,12 +15,19 @@ if [ "$(docker ps -q -f name=nginx)" ]; then
     fi
 fi
 
+if [ -z "$1" ]
+  then
+    echo "Please supply the domain name: 'add-domain.sh domain.com'"
+    exit 1
+fi
+
 docker run --rm -i \
--v "${PERSISTENT_ROOT}/letsencrypt:/etc/letsencrypt" \
--v "${PERSISTENT_ROOT}/certbot:/var/www/certbot" \
+-v "${PERSISTENT_ROOT}/docker-data/letsencrypt:/etc/letsencrypt" \
+-v "${PERSISTENT_ROOT}/docker-data/certbot:/var/www/certbot" \
 -p 80:80 \
 -p 443:443 \
-certbot/certbot 'renew' '--standalone'
+certbot/certbot 'certonly' '--standalone' \
+"-d $1" '--agree-tos' "-m $LE_EMAIL"
 
 if [ $DOCKER_STOPPED = true ]; then
     docker container start nginx
